bugfix: don't create dupe bookmarks on fast saves
Fixes #21

Before, if you saved a new bookmark and then selected it and saved
again, you'd create a duplicate. And there would be no delete button.
This is because there was no id associated with it until the refresh
loop. This is because I forgot to attach the id in the response to the
POST that saves the bookmark.