Sqlite for bookmarks. Stop conflicts, show in UI.
Bookmarks are graduating ahead of schedule from a text file to a real
database, sqlite3. While that's the major change here, I'm doing this
for the minor change, which is that I use a version field to prevent two
different users from saving the same bookmark based on an outdated
version written by somebody else.

As a more minor change, I added error messages for saving and deleting
bookmarks.