Bookmarks: line wraping for titles, more buttons
Annoyngly, if we add line wrap to the bookmark titles for read only, we
have to add it for editors. Editors with line wrap? In an editor without
newlines? Didn't feel like figuring that out. Instead, I made it so that
even editors see the line-wrapped title (which is now in a div) and have
to hit an edit button to start editing it. So now it's even more
complicated. Oh well.