Load geojson via javascript again, but cleaner
Before it was two files. I combined it into one. Hopefully that's faster
to download. But also it'll all show up as one instead of the US showing
up first. This does mean that the US could show up later though, so I
tried adding a loading screen. We'll see how that all goes.

Also, unlike before, I'm having nginx send the thing gzipped. That
should be a little faster.