Implement internal endpoint for getting tutorial type
It's an internal endpoint because tutorial type is gotten from
javascript as part of the general status endpoint. However I'd like to
factor out the tutorial-getting stuff from Python to Go. We can call
this Go endpoint thus from Python. Eventually this logic will be part of
the overall status endpoint again, this time on the Go side.